These files are a start to my first website.
I'm not sure what I plan on doing yet but I'll add more to it as I go on. I have a JavaScript file linked to this project but for now since there is no text in that .js file I decided to hold off on uploading it. Since its blank...